requirements.txt
flask>=3.0,<4
requests>=2.32,<3
python-dotenv>=1.0,<2
